requirements.txt
boto3==1.12.5
botocore==1.15.5
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
docutils==0.15.2
future==0.18.2
jmespath==0.9.4
pycparser==2.19
python-dateutil==2.8.1
python-telegram-bot==12.4.2
s3transfer==0.3.3
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
